+++ caglioti.h
@@ -2,7 +2,7 @@
 #define LIBCAGLIOTI_H
 
 // One include to rule the all!!
-#include "libcaglioti/utils.h"
-#include "libcaglioti/listaseq.h"
+#include "lib/utils.h"
+#include "lib/listaseq.h"
 
 #endif
