+++ lib/listaseq.c
