+++ lib/listaseq.h
